Refactor the person class to a separate directory
