+++ people/corrector.rb
