+++ people/student.rb
@@ -1,5 +1,5 @@
 # rubocop:disable Style/OptionalBooleanParameter
-require './person'
+require_relative 'person'
 
 class Student < Person
   attr_reader :classroom
