+++ people/teacher.rb
@@ -1,4 +1,4 @@
-require './person'
+require_relative 'person'
 
 class Teacher < Person
   def initialize(age, specialization, name = 'Unknown', parent_permission: true)
